requirements.txt
black==24.4.2
click==8.1.7
colorama==0.4.6
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.1
pygame-ce==2.4.1
